renamed compare to prepare_for_union
